+++ src/HeadUnit/HeadUnitStubImpl.hpp
@@ -20,7 +20,7 @@ public:
     virtual void setSensorRpm(const std::shared_ptr<CommonAPI::ClientId> _client, uint16_t _sensorRpm, setSensorRpmReply_t _return);
 
     // Override the setSteering function to handle steering value updates
-    virtual void setSteering(const std::shared_ptr<CommonAPI::ClientId> _client, uint16_t _steering, setSteeringReply_t _return);
+    virtual void setSteering(const std::shared_ptr<CommonAPI::ClientId> _client, double _steering, setSteeringReply_t _return);
 };
 
 #endif
